RENAME smac.smbo to smac.optimizer
