+++ test/test_initial_design/test_single_config_initial_design.py
@@ -13,7 +13,7 @@ from smac.stats.stats import Stats
 from smac.tae.execute_func import ExecuteTAFuncDict
 from smac.initial_design.default_configuration_design import DefaultConfiguration
 from smac.utils.io.traj_logging import TrajLogger
-from smac.smbo.objective import average_cost
+from smac.optimizer.objective import average_cost
 
 
 class TestSingleInitialDesign(unittest.TestCase):
