+++ test/test_integration/eips.py
@@ -3,11 +3,11 @@ import unittest
 
 from smac.scenario.scenario import Scenario
 from smac.utils import test_helpers
-from smac.smbo.smbo import SMBO, get_types
+from smac.optimizer.smbo import SMBO, get_types
 from smac.runhistory.runhistory2epm import RunHistory2EPM4EIPS
 from smac.epm.uncorrelated_mo_rf_with_instances import \
     UncorrelatedMultiObjectiveRandomForestWithInstances
-from smac.smbo.acquisition import EIPS
+from smac.optimizer.acquisition import EIPS
 from smac.tae.execute_func import ExecuteTAFunc
