+++ test/test_scenario/test_scenario.py
@@ -20,7 +20,7 @@ from smac.scenario.scenario import Scenario, _is_truthy
 from smac.configspace import ConfigurationSpace
 from smac.utils.merge_foreign_data import merge_foreign_data
 from smac.runhistory.runhistory import RunHistory
-from smac.smbo.objective import average_cost
+from smac.optimizer.objective import average_cost
 from smac.tae.execute_ta_run import StatusType
 
 if sys.version_info[0] == 2:
