+++ test/test_smbo/test_pSMAC.py
@@ -6,8 +6,8 @@ import unittest
 from smac.runhistory.runhistory import RunHistory, RunKey
 from smac.utils import test_helpers
 from smac.tae.execute_ta_run import StatusType
-from smac.smbo import pSMAC
-from smac.smbo.objective import average_cost
+from smac.optimizer import pSMAC
+from smac.optimizer.objective import average_cost
 
 
 class TestPSMAC(unittest.TestCase):
